Clean up attribute folders
