+++ client/anoixo-client/src/texts/NLF/InflectedFormField.tsx
@@ -1,15 +1,15 @@
 import React, { memo, ChangeEvent, useCallback } from 'react';
 import transliteratedMatchSorter from './utils/GreekMatchSorter';
-import { useAttributeQueryCache } from '../../../query/AttributeQueryCache';
+import { useAttributeQueryCache } from '../../query/AttributeQueryCache';
 import { useUID } from 'react-uid';
-import { AttributeComponent, AttributeComponentProps } from '../../../query/Attributes/AttributeComponent';
-import AttributeEditor from '../../../query/Attributes/AttributeEditor';
+import { AttributeComponent, AttributeComponentProps } from '../../query/Attributes/AttributeComponent';
+import AttributeEditor from '../../query/Attributes/AttributeEditor';
 import Autocomplete from '@material-ui/lab/Autocomplete';
 import TextField from '@material-ui/core/TextField';
 import { FilterOptionsState } from '@material-ui/lab';
-import './LexicalFormField.css';
+import './InflectedFormField.css';
 
-const LexicalFormField: AttributeComponent = memo((props: AttributeComponentProps) => {
+const InflectedFormField: AttributeComponent = memo((props: AttributeComponentProps) => {
     const uid = 'attr-field-' + useUID();
     const { updateAttr, id } = props;
     const onChange = useCallback(
@@ -20,13 +20,13 @@ const LexicalFormField: AttributeComponent = memo((props: AttributeComponentProp
       [updateAttr, id]
     );
 
-    const allLexicalForms = useAttributeQueryCache('lemma');
+    const allInflectedForms = useAttributeQueryCache('normalized');
 
-    const filterLexicalForms = (lexicalForms: string[], state: FilterOptionsState<string>) => {
-      return transliteratedMatchSorter(lexicalForms, state.inputValue, 8);
-    }
+    const filterInflectedForms = (inflectedForms: string[], state: FilterOptionsState<string>) => {
+      return transliteratedMatchSorter(inflectedForms, state.inputValue, 8);
+    };
 
-    const label = 'Lexical Form';
+    const label = 'Inflected Form';
 
     return (
       <AttributeEditor labelText={label} labelProps={{ htmlFor: uid }} enabled={props.enabled}>
@@ -34,10 +34,10 @@ const LexicalFormField: AttributeComponent = memo((props: AttributeComponentProp
           className='attribute-input'
           id={uid}
           renderInput={
-            (params) => <TextField {...params} className='lexical-form-selector' placeholder='Any'/>
+            (params) => <TextField {...params} className='inflected-form-selector' placeholder='Any'/>
           }
-          options={allLexicalForms}
-          filterOptions={filterLexicalForms}
+          options={allInflectedForms}
+          filterOptions={filterInflectedForms}
           onChange={onChange}
         />
       </AttributeEditor>
@@ -45,4 +45,4 @@ const LexicalFormField: AttributeComponent = memo((props: AttributeComponentProp
   }
 );
 
-export default LexicalFormField;
+export default InflectedFormField;
